Use stdlib functions for hash key discovery
When fiddling around with server names in the configuration, you'd have
to do a clean sweep of the whole configuration to ensure that there are
no problems with the fact existence and the ordering.

Using dig (and has_key for compatibility reasons to not change anything
front-facing), we can avoid this problem.