Add option to disable ordering dependencies on Openvpn::Client
We have a setup where we manage client certificates outside of puppet
and we'd like to migrate our servers to this module. We are perfectly
fine with managing the certificates with our existing workflow. Since we
do not need it, we do not configure a `Client[$name]` resource for the
matching `Client_specific_config[$name]`, but currently this code results
in an error because it expects such a resource to be defined. Resolve
this by making the ordering dependency optional.

Signed-off-by: Florian Pritz <florian.pritz@rise-world.com>